scripts/requirements.txt
actionlib-msgs==1.13.0
catkin-pkg==0.5.2
control-msgs==1.5.2
diagnostic-msgs==1.13.0
distro==1.7.0
docutils==0.19
geometry-msgs==1.13.0
map-msgs==1.14.0
move-base-msgs==1.14.0
nav-msgs==1.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
roscpp==1.15.7
rosgraph-msgs==1.11.3
rospkg==1.4.0
rospy-all==0.0.1
shape-msgs==1.13.0
six==1.16.0
std-msgs==0.5.13
stereo-msgs==1.13.0
tf2-msgs==0.7.2
trajectory-msgs==1.13.0
visualization-msgs==1.13.0
